+++ src/attribute-name.ts
@@ -1,0 +1,3 @@
+import { BioAttribute } from './types';
+
+export const attributeName = (attr: string|BioAttribute) => typeof attr === 'string' ? attr : attr.name;
