+++ src/is-registered.ts
@@ -1,0 +1,8 @@
+
+export const isRegistered = (name: string) => {
+  switch (document.createElement(name).constructor) {
+    case HTMLElement: return false;
+    case HTMLUnknownElement: return false;
+  }
+  return true;
+};
