+++ src/types.ts
@@ -1,0 +1,5 @@
+
+export interface BioAttribute {
+  name: string;
+  converter: Function;
+}
